chore: Refactoring action execution flows to clean up files
