+++ app/server/appsmith-server/src/main/java/com/appsmith/server/controllers/ActionController.java
@@ -4,6 +4,7 @@ import com.appsmith.server.constants.Url;
 import com.appsmith.server.controllers.ce.ActionControllerCE;
 import com.appsmith.server.services.LayoutActionService;
 import com.appsmith.server.services.NewActionService;
+import com.appsmith.server.solutions.ActionExecutionSolution;
 import com.appsmith.server.solutions.RefactoringSolution;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,9 +17,10 @@ public class ActionController extends ActionControllerCE {
 
     public ActionController(LayoutActionService layoutActionService,
                             NewActionService newActionService,
-                            RefactoringSolution refactoringSolution) {
+                            RefactoringSolution refactoringSolution,
+                            ActionExecutionSolution actionExecutionSolution) {
 
-        super(layoutActionService, newActionService, refactoringSolution);
+        super(layoutActionService, newActionService, refactoringSolution, actionExecutionSolution);
 
     }
