+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/ActionExecutionSolution.java
@@ -1,0 +1,6 @@
+package com.appsmith.server.solutions;
+
+import com.appsmith.server.solutions.ce.ActionExecutionSolutionCE;
+
+public interface ActionExecutionSolution extends ActionExecutionSolutionCE {
+}
