+++ app/server/appsmith-server/src/test/java/com/appsmith/server/services/ce/ActionServiceCE_Test.java
@@ -1,26 +1,16 @@
 package com.appsmith.server.services.ce;
 
-import com.appsmith.external.constants.DisplayDataType;
-import com.appsmith.external.dtos.ExecuteActionDTO;
-import com.appsmith.external.exceptions.pluginExceptions.AppsmithPluginError;
-import com.appsmith.external.exceptions.pluginExceptions.AppsmithPluginException;
-import com.appsmith.external.exceptions.pluginExceptions.StaleConnectionException;
 import com.appsmith.external.helpers.AppsmithBeanUtils;
 import com.appsmith.external.helpers.AppsmithEventContext;
 import com.appsmith.external.helpers.AppsmithEventContextType;
 import com.appsmith.external.models.ActionConfiguration;
 import com.appsmith.external.models.ActionDTO;
-import com.appsmith.external.models.ActionExecutionResult;
 import com.appsmith.external.models.Datasource;
 import com.appsmith.external.models.DatasourceConfiguration;
 import com.appsmith.external.models.DefaultResources;
-import com.appsmith.external.models.PaginationField;
 import com.appsmith.external.models.PaginationType;
-import com.appsmith.external.models.ParsedDataType;
 import com.appsmith.external.models.Policy;
 import com.appsmith.external.models.Property;
-import com.appsmith.external.models.WidgetSuggestionDTO;
-import com.appsmith.external.models.WidgetType;
 import com.appsmith.external.plugins.PluginExecutor;
 import com.appsmith.server.acl.AclPermission;
 import com.appsmith.server.constants.FieldName;
@@ -37,13 +27,11 @@ import com.appsmith.server.dtos.ActionViewDTO;
 import com.appsmith.server.dtos.ApplicationAccessDTO;
 import com.appsmith.server.dtos.DslActionDTO;
 import com.appsmith.server.dtos.LayoutDTO;
-import com.appsmith.server.dtos.MockDataSource;
 import com.appsmith.server.dtos.PageDTO;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
 import com.appsmith.server.helpers.MockPluginExecutor;
 import com.appsmith.server.helpers.PluginExecutorHelper;
-import com.appsmith.server.helpers.WidgetSuggestionHelper;
 import com.appsmith.server.repositories.DatasourceRepository;
 import com.appsmith.server.repositories.PermissionGroupRepository;
 import com.appsmith.server.repositories.PluginRepository;
@@ -61,8 +49,6 @@ import com.appsmith.server.services.PluginService;
 import com.appsmith.server.services.UserService;
 import com.appsmith.server.services.WorkspaceService;
 import com.appsmith.server.solutions.ImportExportApplicationService;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import net.minidev.json.JSONArray;
@@ -85,8 +71,6 @@ import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 import reactor.util.function.Tuples;
 
-import java.time.Duration;
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -653,343 +637,7 @@ public class ActionServiceCE_Test {
                 .verify();
     }
 
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testVariableSubstitution() {
-        String json = "{\n" +
-                "  \n" +
-                "  \"deleted\": false,\n" +
-                "  \"config\": {\n" +
-                "    \"CONTAINER_WIDGET\": [\n" +
-                "      {\n" +
-                "        \"_id\": \"7\",\n" +
-                "        \"sectionName\": \"General\",\n" +
-                "        \"children\": [\n" +
-                "          {\n" +
-                "            \"_id\": \"7.1\",\n" +
-                "            \"helpText\": \"Use a html color name, HEX, RGB or RGBA value\",\n" +
-                "            \"placeholderText\": \"#FFFFFF / Gray / rgb(255, 99, 71)\",\n" +
-                "            \"propertyName\": \"backgroundColor\",\n" +
-                "            \"label\": \"Background Color\",\n" +
-                "            \"controlType\": \"INPUT_TEXT\"\n" +
-                "          },\n" +
-                "          {\n" +
-                "            \"_id\": \"7.2\",\n" +
-                "            \"helpText\": \"Controls the visibility of the widget\",\n" +
-                "            \"propertyName\": \"isVisible\",\n" +
-                "            \"label\": \"Visible\",\n" +
-                "            \"controlType\": \"SWITCH\",\n" +
-                "            \"isJSConvertible\": true\n" +
-                "          }\n" +
-                "        ]\n" +
-                "      }\n" +
-                "    ]\n" +
-                "  },\n" +
-                "  \"name\": \"propertyPane\"\n" +
-                "}";
 
-        ActionDTO action = new ActionDTO();
-        action.setActionConfiguration(new ActionConfiguration());
-        action.getActionConfiguration().setBody("{{Input.text}}");
-
-        ActionDTO renderedAction = newActionService.variableSubstitution(action, Map.of("Input.text", json));
-        assertThat(renderedAction).isNotNull();
-        assertThat(renderedAction.getActionConfiguration().getBody()).isEqualTo(json);
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testVariableSubstitutionWithNewline() {
-        ActionDTO action = new ActionDTO();
-        action.setActionConfiguration(new ActionConfiguration());
-        action.getActionConfiguration().setBody("{{Input.text}}");
-
-        ActionDTO renderedAction = newActionService.variableSubstitution(action, Map.of("Input.text", "name\nvalue"));
-        assertThat(renderedAction).isNotNull();
-        assertThat(renderedAction.getActionConfiguration().getBody()).isEqualTo("name\nvalue");
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecute() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult, List.of(new ParsedDataType(DisplayDataType.RAW)));
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteNullRequestBody() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.GET);
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setName("testActionExecuteNullRequestBody");
-        action.setPageId(testPage.getId());
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult, List.of(new ParsedDataType(DisplayDataType.RAW)));
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteDbQuery() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setBody("select * from users");
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecuteDbQuery");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteErrorResponse() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHeaders(List.of(
-                new Property("random-header-key", "random-header-value"),
-                new Property("", "")
-        ));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecuteErrorResponse");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        AppsmithPluginException pluginException = new AppsmithPluginException(AppsmithPluginError.PLUGIN_ERROR);
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.executeParameterizedWithMetrics(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any())).thenReturn(Mono.error(pluginException));
-        Mockito.when(pluginExecutor.datasourceCreate(Mockito.any())).thenReturn(Mono.empty());
-
-        Mono<ActionExecutionResult> executionResultMono = newActionService.executeAction(executeActionDTO, null);
-
-        StepVerifier.create(executionResultMono)
-                .assertNext(result -> {
-                    assertThat(result.getIsExecutionSuccess()).isFalse();
-                    assertThat(result.getStatusCode()).isEqualTo(pluginException.getAppErrorCode().toString());
-                    assertThat(result.getTitle()).isEqualTo(pluginException.getTitle());
-                    assertThat(result.getRequest().getActionId()).isEqualTo(createdAction.getId());
-                    assertThat(result.getRequest().getRequestedAt()).isBefore(Instant.now());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteNullPaginationParameters() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHeaders(List.of(
-                new Property("random-header-key", "random-header-value"),
-                new Property("", "")
-        ));
-        actionConfiguration.setPaginationType(PaginationType.URL);
-        actionConfiguration.setNext(null);
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecuteErrorResponse");
-        DatasourceConfiguration datasourceConfiguration = new DatasourceConfiguration();
-        datasource.setDatasourceConfiguration(datasourceConfiguration);
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-        executeActionDTO.setPaginationField(PaginationField.NEXT);
-
-        AppsmithPluginException pluginException = new AppsmithPluginException(AppsmithPluginError.PLUGIN_ERROR);
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.executeParameterizedWithMetrics(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any())).thenReturn(Mono.error(pluginException));
-        Mockito.when(pluginExecutor.datasourceCreate(Mockito.any())).thenReturn(Mono.empty());
-
-        Mono<ActionExecutionResult> executionResultMono = newActionService.executeAction(executeActionDTO, null);
-
-        StepVerifier.create(executionResultMono)
-                .assertNext(result -> {
-                    assertThat(result.getIsExecutionSuccess()).isFalse();
-                    assertThat(result.getStatusCode()).isEqualTo(pluginException.getAppErrorCode().toString());
-                    assertThat(result.getTitle()).isEqualTo(pluginException.getTitle());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteSecondaryStaleConnection() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHeaders(List.of(
-                new Property("random-header-key", "random-header-value"),
-                new Property("", "")
-        ));
-        actionConfiguration.setTimeoutInMillisecond(String.valueOf(1000));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecuteSecondaryStaleConnection");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.executeParameterizedWithMetrics(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.error(new StaleConnectionException())).thenReturn(Mono.error(new StaleConnectionException()));
-        Mockito.when(pluginExecutor.datasourceCreate(Mockito.any())).thenReturn(Mono.empty());
-
-        Mono<ActionExecutionResult> executionResultMono = newActionService.executeAction(executeActionDTO, null);
-
-        StepVerifier.create(executionResultMono)
-                .assertNext(result -> {
-                    assertThat(result.getIsExecutionSuccess()).isFalse();
-                    assertThat(result.getStatusCode()).isEqualTo(AppsmithPluginError.STALE_CONNECTION_ERROR.getAppErrorCode());
-                    assertThat(result.getTitle()).isEqualTo(AppsmithPluginError.STALE_CONNECTION_ERROR.getTitle());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteTimeout() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHeaders(List.of(
-                new Property("random-header-key", "random-header-value"),
-                new Property("", "")
-        ));
-        actionConfiguration.setTimeoutInMillisecond(String.valueOf(10));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecuteTimeout");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.executeParameterizedWithMetrics(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any()))
-                .thenAnswer(x -> Mono.delay(Duration.ofMillis(1000)).ofType(ActionExecutionResult.class));
-        Mockito.when(pluginExecutor.datasourceCreate(Mockito.any())).thenReturn(Mono.empty());
-
-        Mono<ActionExecutionResult> executionResultMono = newActionService.executeAction(executeActionDTO, null);
-
-        StepVerifier.create(executionResultMono)
-                .assertNext(result -> {
-                    assertThat(result.getIsExecutionSuccess()).isFalse();
-                    assertThat(result.getStatusCode()).isEqualTo(AppsmithPluginError.PLUGIN_QUERY_TIMEOUT_ERROR.getAppErrorCode().toString());
-                    assertThat(result.getTitle()).isEqualTo(AppsmithPluginError.PLUGIN_QUERY_TIMEOUT_ERROR.getTitle());
-                })
-                .verifyComplete();
-    }
 
     @Test
     @WithUserDetails(value = "api_user")
@@ -1055,74 +703,6 @@ public class ActionServiceCE_Test {
 
     @Test
     @WithUserDetails(value = "api_user")
-    public void checkRecoveryFromStaleConnections() {
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.executeParameterizedWithMetrics(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any()))
-                .thenThrow(new StaleConnectionException())
-                .thenReturn(Mono.just(mockResult));
-        Mockito.when(pluginExecutor.datasourceCreate(Mockito.any())).thenReturn(Mono.empty());
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setBody("select * from users");
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("checkRecoveryFromStaleConnections");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        Mono<ActionExecutionResult> actionExecutionResultMono = newActionService.executeAction(executeActionDTO, null);
-
-        StepVerifier.create(actionExecutionResultMono)
-                .assertNext(result -> {
-                    assertThat(result).isNotNull();
-                    assertThat(result.getBody()).isEqualTo(mockResult.getBody());
-                })
-                .verifyComplete();
-    }
-
-    private void executeAndAssertAction(ExecuteActionDTO executeActionDTO, ActionConfiguration actionConfiguration,
-                                        ActionExecutionResult mockResult, List<ParsedDataType> expectedReturnDataTypes) {
-
-        List<WidgetSuggestionDTO> expectedWidgets = mockResult.getSuggestedWidgets();
-        Mono<ActionExecutionResult> actionExecutionResultMono = executeAction(executeActionDTO, actionConfiguration, mockResult);
-
-        StepVerifier.create(actionExecutionResultMono)
-                .assertNext(result -> {
-                    assertThat(result).isNotNull();
-                    assertThat(result.getBody()).isEqualTo(mockResult.getBody());
-                    assertThat(result.getDataTypes()).hasToString(expectedReturnDataTypes.toString());
-                    assertThat(result.getSuggestedWidgets())
-                            .usingRecursiveFieldByFieldElementComparator()
-                            .containsExactlyInAnyOrderElementsOf(expectedWidgets);
-                    assertThat(result.getRequest().getActionId()).isEqualTo(executeActionDTO.getActionId());
-                    assertThat(result.getRequest().getRequestedAt()).isBefore(Instant.now());
-                })
-                .verifyComplete();
-    }
-
-    private Mono<ActionExecutionResult> executeAction(ExecuteActionDTO executeActionDTO, ActionConfiguration actionConfiguration, ActionExecutionResult mockResult) {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.executeParameterizedWithMetrics(Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any(), Mockito.any())).thenReturn(Mono.just(mockResult));
-        Mockito.when(pluginExecutor.datasourceCreate(Mockito.any())).thenReturn(Mono.empty());
-
-        Mono<ActionExecutionResult> actionExecutionResultMono = newActionService.executeAction(executeActionDTO, null);
-        return actionExecutionResultMono;
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
     public void getActionInViewMode() {
         Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(new MockPluginExecutor()));
 
@@ -1154,48 +734,6 @@ public class ActionServiceCE_Test {
                 .verifyComplete();
     }
 
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void executeActionWithExternalDatasource() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(new MockPluginExecutor()));
-        Mockito.when(pluginService.getEditorConfigLabelMap(Mockito.anyString())).thenReturn(Mono.just(new HashMap<>()));
-
-        Datasource externalDatasource = new Datasource();
-        externalDatasource.setName("Default Database");
-        externalDatasource.setWorkspaceId(workspaceId);
-        Plugin restApiPlugin = pluginRepository.findByPackageName("restapi-plugin").block();
-        externalDatasource.setPluginId(restApiPlugin.getId());
-        DatasourceConfiguration datasourceConfiguration = new DatasourceConfiguration();
-        datasourceConfiguration.setUrl("some url here");
-        externalDatasource.setDatasourceConfiguration(datasourceConfiguration);
-        Datasource savedDs = datasourceService.create(externalDatasource).block();
-
-        ActionDTO action = new ActionDTO();
-        action.setName("actionWithExternalDatasource");
-        action.setPageId(testPage.getId());
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.GET);
-        action.setActionConfiguration(actionConfiguration);
-        action.setDatasource(savedDs);
-
-
-        Mono<ActionExecutionResult> resultMono = layoutActionService.createSingleAction(action, Boolean.FALSE)
-                .flatMap(savedAction -> {
-                    ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-                    executeActionDTO.setActionId(savedAction.getId());
-                    executeActionDTO.setViewMode(false);
-                    return newActionService.executeAction(executeActionDTO, null);
-                });
-
-
-        StepVerifier
-                .create(resultMono)
-                .assertNext(result -> {
-                    assertThat(result).isNotNull();
-                    assertThat(result.getStatusCode()).isEqualTo("200");
-                })
-                .verifyComplete();
-    }
 
     @Test
     @WithUserDetails(value = "api_user")
@@ -1547,1108 +1085,6 @@ public class ActionServiceCE_Test {
                 .verifyComplete();
     }
 
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteWithTableReturnType() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("[\n" +
-                "{\"name\": \"Richard\", \"profession\": \"medical\"},\n" +
-                "{\"name\": \"John\", \"profession\": \"self employed\"},\n" +
-                "{\"name\": \"Mary\", \"profession\": \"engineer\"}\n" +
-                "]");
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.TABLE), new ParsedDataType(DisplayDataType.JSON)
-                        , new ParsedDataType(DisplayDataType.RAW)));
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteWithJsonReturnType() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("{\n" +
-                "  \"name\":\"John\",\n" +
-                "  \"age\":30,\n" +
-                "  \"cars\": {\n" +
-                "    \"car1\":\"Ford\",\n" +
-                "    \"car2\":\"BMW\",\n" +
-                "    \"car3\":\"Fiat\"\n" +
-                "  }\n" +
-                " }");
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.JSON), new ParsedDataType(DisplayDataType.RAW)));
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteWithPreAssignedReturnType() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("{\n" +
-                "  \"name\":\"John\",\n" +
-                "  \"age\":30,\n" +
-                "  \"cars\": {\n" +
-                "    \"car1\":\"Ford\",\n" +
-                "    \"car2\":\"BMW\",\n" +
-                "    \"car3\":\"Fiat\"\n" +
-                "  }\n" +
-                " }");
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testActionExecuteReturnTypeWithNullResultBody() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(null);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult, new ArrayList<>());
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithChartWidgetData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\": [\n" +
-                "  {\n" +
-                "    \"x\": \"Mon\",\n" +
-                "    \"y\": 10000\n" +
-                "  },\n" +
-                "  {\n" +
-                "    \"x\": \"Tue\",\n" +
-                "    \"y\": 12000\n" +
-                "  },\n" +
-                "  {\n" +
-                "    \"x\": \"Wed\",\n" +
-                "    \"y\": 32000\n" +
-                "  },\n" +
-                "  {\n" +
-                "    \"x\": \"Thu\",\n" +
-                "    \"y\": 28000\n" +
-                "  },\n" +
-                "  {\n" +
-                "    \"x\": \"Fri\",\n" +
-                "    \"y\": 14000\n" +
-                "  },\n" +
-                "  {\n" +
-                "    \"x\": \"Sat\",\n" +
-                "    \"y\": 19000\n" +
-                "  },\n" +
-                "  {\n" +
-                "    \"x\": \"Sun\",\n" +
-                "    \"y\": 36000\n" +
-                "  }\n" +
-                "]}";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.CHART_WIDGET, "x", "y"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.SELECT_WIDGET, "x", "x"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TABLE_WIDGET_V2));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithTableWidgetData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\": [\n" +
-                "\t{\n" +
-                "\t\t\"id\": \"0001\",\n" +
-                "\t\t\"type\": \"donut\",\n" +
-                "\t\t\"name\": \"Cake\",\n" +
-                "\t\t\"ppu\": 0.55,\n" +
-                "\t\t\"batters\":\n" +
-                "\t\t\t{\n" +
-                "\t\t\t\t\"batter\":\n" +
-                "\t\t\t\t\t[\n" +
-                "\t\t\t\t\t\t{ \"id\": \"1001\", \"type\": \"Regular\" },\n" +
-                "\t\t\t\t\t\t{ \"id\": \"1002\", \"type\": \"Chocolate\" },\n" +
-                "\t\t\t\t\t\t{ \"id\": \"1003\", \"type\": \"Blueberry\" },\n" +
-                "\t\t\t\t\t\t{ \"id\": \"1004\", \"type\": \"Devil's Food\" }\n" +
-                "\t\t\t\t\t]\n" +
-                "\t\t\t},\n" +
-                "\t\t\"topping\":\n" +
-                "\t\t\t[\n" +
-                "\t\t\t\t{ \"id\": \"5001\", \"type\": \"None\" },\n" +
-                "\t\t\t\t{ \"id\": \"5002\", \"type\": \"Glazed\" },\n" +
-                "\t\t\t\t{ \"id\": \"5005\", \"type\": \"Sugar\" },\n" +
-                "\t\t\t\t{ \"id\": \"5007\", \"type\": \"Powdered Sugar\" },\n" +
-                "\t\t\t\t{ \"id\": \"5006\", \"type\": \"Chocolate with Sprinkles\" },\n" +
-                "\t\t\t\t{ \"id\": \"5003\", \"type\": \"Chocolate\" },\n" +
-                "\t\t\t\t{ \"id\": \"5004\", \"type\": \"Maple\" }\n" +
-                "\t\t\t]\n" +
-                "\t},\n" +
-                "\t{\n" +
-                "\t\t\"id\": \"0002\",\n" +
-                "\t\t\"type\": \"donut\",\n" +
-                "\t\t\"name\": \"Raised\",\n" +
-                "\t\t\"ppu\": 0.55,\n" +
-                "\t\t\"batters\":\n" +
-                "\t\t\t{\n" +
-                "\t\t\t\t\"batter\":\n" +
-                "\t\t\t\t\t[\n" +
-                "\t\t\t\t\t\t{ \"id\": \"1001\", \"type\": \"Regular\" }\n" +
-                "\t\t\t\t\t]\n" +
-                "\t\t\t},\n" +
-                "\t\t\"topping\":\n" +
-                "\t\t\t[\n" +
-                "\t\t\t\t{ \"id\": \"5001\", \"type\": \"None\" },\n" +
-                "\t\t\t\t{ \"id\": \"5002\", \"type\": \"Glazed\" },\n" +
-                "\t\t\t\t{ \"id\": \"5005\", \"type\": \"Sugar\" },\n" +
-                "\t\t\t\t{ \"id\": \"5003\", \"type\": \"Chocolate\" },\n" +
-                "\t\t\t\t{ \"id\": \"5004\", \"type\": \"Maple\" }\n" +
-                "\t\t\t]\n" +
-                "\t},\n" +
-                "\t{\n" +
-                "\t\t\"id\": \"0003\",\n" +
-                "\t\t\"type\": \"donut\",\n" +
-                "\t\t\"name\": \"Old Fashioned\",\n" +
-                "\t\t\"ppu\": 0.55,\n" +
-                "\t\t\"batters\":\n" +
-                "\t\t\t{\n" +
-                "\t\t\t\t\"batter\":\n" +
-                "\t\t\t\t\t[\n" +
-                "\t\t\t\t\t\t{ \"id\": \"1001\", \"type\": \"Regular\" },\n" +
-                "\t\t\t\t\t\t{ \"id\": \"1002\", \"type\": \"Chocolate\" }\n" +
-                "\t\t\t\t\t]\n" +
-                "\t\t\t},\n" +
-                "\t\t\"topping\":\n" +
-                "\t\t\t[\n" +
-                "\t\t\t\t{ \"id\": \"5001\", \"type\": \"None\" },\n" +
-                "\t\t\t\t{ \"id\": \"5002\", \"type\": \"Glazed\" },\n" +
-                "\t\t\t\t{ \"id\": \"5003\", \"type\": \"Chocolate\" },\n" +
-                "\t\t\t\t{ \"id\": \"5004\", \"type\": \"Maple\" }\n" +
-                "\t\t\t]\n" +
-                "\t}\n" +
-                "]}";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.CHART_WIDGET, "id", "ppu"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.SELECT_WIDGET, "id", "type"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TABLE_WIDGET_V2));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithListWidgetData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\": [\n" +
-                "    {\n" +
-                "        \"url\": \"images/thumbnails/0001.jpg\",\n" +
-                "        \"width\": 32,\n" +
-                "        \"height\": 32\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0001.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0002.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0002.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0003.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0004.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0005.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0006.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0007.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0008.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0009.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    },\n" +
-                "    {\n" +
-                "        \"url\": \"images/0010.jpg\",\n" +
-                "        \"width\": 200,\n" +
-                "        \"height\": 200\n" +
-                "    }\n" +
-                "]}";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.CHART_WIDGET, "url", "width"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.SELECT_WIDGET, "url", "url"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TABLE_WIDGET_V2));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithDropdownWidgetData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\": [\n" +
-                "    {\n" +
-                "     \"CarType\": \"BMW\",\n" +
-                "     \"carID\": \"bmw123\"\n" +
-                "     },\n" +
-                "      {\n" +
-                "     \"CarType\": \"mercedes\",\n" +
-                "     \"carID\": \"merc123\"\n" +
-                "      },\n" +
-                "      {\n" +
-                "     \"CarType\": \"volvo\",\n" +
-                "     \"carID\": \"vol123r\"\n" +
-                "       },\n" +
-                "       {\n" +
-                "     \"CarType\": \"ford\",\n" +
-                "     \"carID\": \"ford123\"\n" +
-                "       }\n" +
-                "  ]}";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.SELECT_WIDGET, "CarType", "carID"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TABLE_WIDGET_V2));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithArrayOfStringsDropDownWidget() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\":[\"string1\", \"string2\", \"string3\", \"string4\"] }";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.INPUT_WIDGET));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithArrayOfArray() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\":[[\"string1\", \"string2\", \"string3\", \"string4\"]," +
-                "[\"string5\", \"string6\", \"string7\", \"string8\"]," +
-                "[\"string9\", \"string10\", \"string11\", \"string12\"]] }";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TABLE_WIDGET_V2));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithEmptyData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\":[] }";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithNumericData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\": [1] }";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.INPUT_WIDGET));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithJsonNodeData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{\"data\": {\n" +
-                "    \"next\": \"https://mock-api.appsmith.com/users?page=2&pageSize=10\",\n" +
-                "    \"previous\": null,\n" +
-                "    \"users\": [\n" +
-                "        {\n" +
-                "            \"id\": 3,\n" +
-                "            \"name\": \"Demetre\",\n" +
-                "            \"status\": \"APPROVED\",\n" +
-                "            \"gender\": \"Male\",\n" +
-                "            \"avatar\": \"https://robohash.org/iustooptiocum.jpg?size=100x100&set=set1\",\n" +
-                "            \"email\": \"aaaa@bbb.com\",\n" +
-                "            \"address\": \"262 Saint Paul Park\",\n" +
-                "            \"createdAt\": \"2020-05-01T17:30:50.000Z\",\n" +
-                "            \"updatedAt\": \"2019-10-08T14:55:53.000Z\"\n" +
-                "        },\n" +
-                "        {\n" +
-                "            \"id\": 4,\n" +
-                "            \"name\": \"Currey\",\n" +
-                "            \"status\": \"APPROVED\",\n" +
-                "            \"gender\": \"Female\",\n" +
-                "            \"avatar\": \"https://robohash.org/aspernaturnatusrepellat.jpg?size=100x100&set=set1\",\n" +
-                "            \"email\": \"cbrayson3@taobao.com\",\n" +
-                "            \"address\": \"35180 Lotheville Street!\",\n" +
-                "            \"createdAt\": \"2019-12-30T03:54:23.000Z\",\n" +
-                "            \"updatedAt\": \"2020-08-12T17:43:01.016Z\"\n" +
-                "        }\n" +
-                "    ]\n" +
-                "}}";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidgetNestedData(WidgetType.TEXT_WIDGET, "users"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidgetNestedData(WidgetType.CHART_WIDGET, "users", "name", "id"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidgetNestedData(WidgetType.TABLE_WIDGET_V2, "users"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidgetNestedData(WidgetType.SELECT_WIDGET, "users", "name", "status"));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithJsonObjectData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\": {\n" +
-                "            \"id\": 1,\n" +
-                "            \"name\": \"Barty Crouch\",\n" +
-                "            \"status\": \"APPROVED\",\n" +
-                "            \"gender\": \"\",\n" +
-                "            \"avatar\": \"https://robohash.org/sednecessitatibuset.png?size=100x100&set=set1\",\n" +
-                "            \"email\": \"barty.crouch@gmail.com\",\n" +
-                "            \"address\": \"St Petersberg #911 4th main\",\n" +
-                "            \"createdAt\": \"2020-03-16T18:00:05.000Z\",\n" +
-                "            \"updatedAt\": \"2020-08-12T17:29:31.980Z\"\n" +
-                "        } }";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionAfterExecutionWithJsonArrayObjectData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{ \"data\": {\n" +
-                "            \"id\": 1,\n" +
-                "            \"name\": \"Barty Crouch\",\n" +
-                "            \"status\": \"APPROVED\",\n" +
-                "            \"gender\": \"\",\n" +
-                "            \"avatar\": \"https://robohash.org/sednecessitatibuset.png?size=100x100&set=set1\",\n" +
-                "            \"email\": \"barty.crouch@gmail.com\",\n" +
-                "            \"address\": \"St Petersberg #911 4th main\",\n" +
-                "            \"createdAt\": \"2020-03-16T18:00:05.000Z\",\n" +
-                "            \"updatedAt\": \"2020-08-12T17:29:31.980Z\"\n" +
-                "        }," +
-                "\"data\": {\n" +
-                "            \"id\": 2,\n" +
-                "            \"name\": \"Jenelle Kibbys\",\n" +
-                "            \"status\": \"APPROVED\",\n" +
-                "            \"gender\": \"Female\",\n" +
-                "            \"avatar\": \"https://robohash.org/quiaasperiorespariatur.bmp?size=100x100&set=set1\",\n" +
-                "            \"email\": \"jkibby1@hp.com\",\n" +
-                "            \"address\": \"85 Tennessee Plaza\",\n" +
-                "            \"createdAt\": \"2019-10-04T03:22:23.000Z\",\n" +
-                "            \"updatedAt\": \"2019-09-11T20:18:38.000Z\"\n" +
-                "        } }";
-        final JsonNode arrNode = new ObjectMapper().readTree(data);
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionNestedData() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{\"data\": {\n" +
-                "    \"next\": \"https://mock-api.appsmith.com/users?page=2&pageSize=10\",\n" +
-                "    \"previous\": null,\n" +
-                "    \"users\": [1, 2, 3]\n" +
-                "}}";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidgetNestedData(WidgetType.TEXT_WIDGET, "users"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidgetNestedData(WidgetType.TABLE_WIDGET_V2, "users"));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void testWidgetSuggestionNestedDataEmpty() throws JsonProcessingException {
-
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        final String data = "{\"data\": {\n" +
-                "    \"next\": \"https://mock-api.appsmith.com/users?page=2&pageSize=10\",\n" +
-                "    \"previous\": null,\n" +
-                "    \"users\": []\n" +
-                "}}";
-        final JsonNode arrNode = new ObjectMapper().readTree(data).get("data");
-        ;
-
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(arrNode);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidgetNestedData(WidgetType.TEXT_WIDGET, null));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void suggestWidget_ArrayListData_SuggestTableTextChartDropDownWidget() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        ArrayList<JSONObject> listData = new ArrayList<>();
-        JSONObject jsonObject = new JSONObject(Map.of("url", "images/thumbnails/0001.jpg", "width", 32, "height", 32));
-        listData.add(jsonObject);
-        jsonObject = new JSONObject(Map.of("url", "images/0001.jpg", "width", 42, "height", 22));
-        listData.add(jsonObject);
-        jsonObject = new JSONObject(Map.of("url", "images/0002.jpg", "width", 52, "height", 12));
-        listData.add(jsonObject);
-        jsonObject = new JSONObject(Map.of("url", "images/0003.jpg", "width", 62, "height", 52));
-        listData.add(jsonObject);
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(listData);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.CHART_WIDGET, "url", "width"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.SELECT_WIDGET, "url", "url"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TABLE_WIDGET_V2));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void suggestWidget_ArrayListData_SuggestTableTextDropDownWidget() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        ArrayList<JSONObject> listData = new ArrayList<>();
-        JSONObject jsonObject = new JSONObject(Map.of("url", "images/thumbnails/0001.jpg", "width", "32", "height", "32"));
-        listData.add(jsonObject);
-        jsonObject = new JSONObject(Map.of("url", "images/0001.jpg", "width", "42", "height", "22"));
-        listData.add(jsonObject);
-        jsonObject = new JSONObject(Map.of("url", "images/0002.jpg", "width", "52", "height", "12"));
-        listData.add(jsonObject);
-        jsonObject = new JSONObject(Map.of("url", "images/0003.jpg", "width", "62", "height", "52"));
-        listData.add(jsonObject);
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(listData);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.SELECT_WIDGET, "width", "url"));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TABLE_WIDGET_V2));
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void suggestWidget_ArrayListDataEmpty_SuggestTextWidget() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(pluginExecutor));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        ArrayList<JSONObject> listData = new ArrayList<>();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody(listData);
-        mockResult.setStatusCode("200");
-        mockResult.setHeaders(objectMapper.valueToTree(Map.of("response-header-key", "response-header-value")));
-        mockResult.setDataTypes(List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setHttpMethod(HttpMethod.POST);
-        actionConfiguration.setBody("random-request-body");
-        actionConfiguration.setHeaders(List.of(new Property("random-header-key", "random-header-value")));
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecute");
-        action.setDatasource(datasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
-
-    }
 
     private Mono<PageDTO> createPage(Application app, PageDTO page) {
         return newPageService
@@ -2787,47 +1223,6 @@ public class ActionServiceCE_Test {
                     assertThat(datasource1.getInvalids()).isEmpty();
                 })
                 .verifyComplete();
-    }
-
-    @Test
-    @WithUserDetails(value = "api_user")
-    public void executeAction_actionOnMockDatasource_success() {
-        Mockito.when(pluginExecutorHelper.getPluginExecutor(Mockito.any())).thenReturn(Mono.just(new MockPluginExecutor()));
-        Mockito.when(pluginService.getEditorConfigLabelMap(Mockito.anyString())).thenReturn(Mono.just(new HashMap<>()));
-        Mockito.when(pluginExecutor.getHintMessages(Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.zip(Mono.just(new HashSet<>()), Mono.just(new HashSet<>())));
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("response-body");
-
-        Plugin installed_plugin = pluginRepository.findByPackageName("restapi-plugin").block();
-        MockDataSource mockDataSource = new MockDataSource();
-        mockDataSource.setName("Users");
-        mockDataSource.setWorkspaceId(workspaceId);
-        mockDataSource.setPackageName("postgres-plugin");
-        mockDataSource.setPluginId(installed_plugin.getId());
-        Datasource mockDatasource = mockDataService.createMockDataSet(mockDataSource).block();
-
-        List<WidgetSuggestionDTO> widgetTypeList = new ArrayList<>();
-        widgetTypeList.add(WidgetSuggestionHelper.getWidget(WidgetType.TEXT_WIDGET));
-        mockResult.setSuggestedWidgets(widgetTypeList);
-
-        ActionDTO action = new ActionDTO();
-        ActionConfiguration actionConfiguration = new ActionConfiguration();
-        actionConfiguration.setBody("select * from users");
-        action.setActionConfiguration(actionConfiguration);
-        action.setPageId(testPage.getId());
-        action.setName("testActionExecuteDbQuery");
-        action.setDatasource(mockDatasource);
-        ActionDTO createdAction = layoutActionService.createSingleAction(action, Boolean.FALSE).block();
-
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId(createdAction.getId());
-        executeActionDTO.setViewMode(false);
-
-        executeAndAssertAction(executeActionDTO, actionConfiguration, mockResult,
-                List.of(new ParsedDataType(DisplayDataType.RAW)));
     }
 
     @Test
