+++ app/server/appsmith-server/src/test/java/com/appsmith/server/services/ce/NewActionServiceCEImplTest.java
@@ -142,10 +142,6 @@ public class NewActionServiceCEImplTest {
     @MockBean
     ObservationRegistry observationRegistry;
 
-    private BodyExtractor.Context context;
-
-    private Map<String, Object> hints;
-
     @BeforeEach
     public void setup() {
         newActionService = new NewActionServiceCEImpl(scheduler,
@@ -156,15 +152,12 @@ public class NewActionServiceCEImplTest {
                 analyticsService,
                 datasourceService,
                 pluginService,
-                datasourceContextService,
                 pluginExecutorHelper,
                 marketplaceService,
                 policyGenerator,
                 newPageService,
                 applicationService,
-                sessionUserService,
                 policyUtils,
-                authenticationValidator,
                 configService,
                 responseUtils,
                 permissionGroupService,
@@ -176,98 +169,6 @@ public class NewActionServiceCEImplTest {
 
         ObservationRegistry.ObservationConfig mockObservationConfig = Mockito.mock(ObservationRegistry.ObservationConfig.class);
         Mockito.when(observationRegistry.observationConfig()).thenReturn(mockObservationConfig);
-    }
-
-    @BeforeEach
-    public void createContext() {
-        final List<HttpMessageReader<?>> messageReaders = new ArrayList<>();
-        messageReaders.add(new DecoderHttpMessageReader<>(new ByteBufferDecoder()));
-        messageReaders.add(new DecoderHttpMessageReader<>(StringDecoder.allMimeTypes()));
-        messageReaders.add(new DecoderHttpMessageReader<>(new Jaxb2XmlDecoder()));
-        messageReaders.add(new DecoderHttpMessageReader<>(new Jackson2JsonDecoder()));
-        messageReaders.add(new FormHttpMessageReader());
-        DefaultPartHttpMessageReader partReader = new DefaultPartHttpMessageReader();
-        messageReaders.add(partReader);
-        messageReaders.add(new MultipartHttpMessageReader(partReader));
-
-        this.context = new BodyExtractor.Context() {
-            @Override
-            public List<HttpMessageReader<?>> messageReaders() {
-                return messageReaders;
-            }
-
-            @Override
-            public Optional<ServerHttpResponse> serverResponse() {
-                return Optional.empty();
-            }
-
-            @Override
-            public Map<String, Object> hints() {
-                return hints;
-            }
-        };
-        this.hints = new HashMap<>();
-    }
-
-    @Test
-    public void testExecuteAction_withoutExecuteActionDTOPart_failsValidation() {
-        final Mono<ActionExecutionResult> actionExecutionResultMono = newActionService.executeAction(Flux.empty(), null, null);
-
-        StepVerifier
-                .create(actionExecutionResultMono)
-                .expectErrorMatches(e -> e instanceof AppsmithException &&
-                        e.getMessage().equals(AppsmithError.INVALID_PARAMETER.getMessage(FieldName.ACTION_ID)))
-                .verify();
-    }
-
-    @Test
-    public void testExecuteAction_withMalformedExecuteActionDTO_failsValidation() {
-        MockServerHttpRequest mock = MockServerHttpRequest
-                .method(HttpMethod.POST, URI.create("https://example.com"))
-                .contentType(new MediaType("multipart", "form-data", Map.of("boundary", "boundary")))
-                .body("""
-                        --boundary\r
-                        Content-Disposition: form-data; name="executeActionDTO"\r
-                        \r
-                        irrelevant content\r
-                        --boundary--\r
-                        """);
-
-        final Flux<Part> partsFlux = BodyExtractors.toParts()
-                .extract(mock, this.context);
-
-        final Mono<ActionExecutionResult> actionExecutionResultMono = newActionService.executeAction(partsFlux, null, null);
-
-        StepVerifier
-                .create(actionExecutionResultMono)
-                .expectErrorMatches(e -> e instanceof AppsmithException &&
-                        e.getMessage().equals(AppsmithError.INVALID_PARAMETER.getMessage("executeActionDTO")))
-                .verify();
-    }
-
-    @Test
-    public void testExecuteAction_withoutActionId_failsValidation() {
-        MockServerHttpRequest mock = MockServerHttpRequest
-                .method(HttpMethod.POST, URI.create("https://example.com"))
-                .contentType(new MediaType("multipart", "form-data", Map.of("boundary", "boundary")))
-                .body("""
-                        --boundary\r
-                        Content-Disposition: form-data; name="executeActionDTO"\r
-                        \r
-                        {"viewMode":false}\r
-                        --boundary--\r
-                        """);
-
-        final Flux<Part> partsFlux = BodyExtractors.toParts()
-                .extract(mock, this.context);
-
-        final Mono<ActionExecutionResult> actionExecutionResultMono = newActionService.executeAction(partsFlux, null, null);
-
-        StepVerifier
-                .create(actionExecutionResultMono)
-                .expectErrorMatches(e -> e instanceof AppsmithException &&
-                        e.getMessage().equals(AppsmithError.INVALID_PARAMETER.getMessage(FieldName.ACTION_ID)))
-                .verify();
     }
 
     @Test
@@ -324,176 +225,4 @@ public class NewActionServiceCEImplTest {
                 .verifyComplete();
     }
 
-    @Test
-    public void testExecuteAPIWithUsualOrderingOfTheParts() {
-        String usualOrderOfParts = """
-                --boundary\r
-                Content-Disposition: form-data; name="executeActionDTO"\r
-                \r
-                {"actionId":"63285a3388e48972c7519b18","viewMode":false,"paramProperties":{"k0":{"datatype": "string"}}}\r
-                --boundary\r
-                Content-Disposition: form-data; name="parameterMap"\r
-                \r
-                {"Input1.text":"k0"}\r
-                --boundary\r
-                Content-Disposition: form-data; name="k0"; filename="blob"\r
-                Content-Type: text/plain\r
-                \r
-                xyz\r
-                --boundary--""";
-
-        MockServerHttpRequest mock = MockServerHttpRequest
-                .method(HttpMethod.POST, URI.create("https://example.com"))
-                .contentType(new MediaType("multipart", "form-data", Map.of("boundary", "boundary")))
-                .body(usualOrderOfParts);
-
-        final Flux<Part> partsFlux = BodyExtractors.toParts()
-                .extract(mock, this.context);
-
-        NewActionServiceCE newActionServiceSpy = spy(newActionService);
-
-        Mono<ActionExecutionResult> actionExecutionResultMono = newActionServiceSpy.executeAction(partsFlux, null, null);
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("test body");
-        mockResult.setTitle("test title");
-
-        NewAction newAction = new NewAction();
-        newAction.setId("63285a3388e48972c7519b18");
-        doReturn(Mono.just(mockResult)).when(newActionServiceSpy).executeAction(any(), any());
-        doReturn(Mono.just(newAction)).when(newActionServiceSpy).findByBranchNameAndDefaultActionId(any(), any(), any());
-
-
-        StepVerifier
-                .create(actionExecutionResultMono)
-                .assertNext(response -> {
-                    assertNotNull(response);
-                    assertTrue(response.getIsExecutionSuccess());
-                    assertEquals(mockResult.getBody().toString(), response.getBody().toString());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    public void testExecuteAPIWithParameterMapAsLastPart() {
-        String parameterMapAtLast = """
-                --boundary\r
-                Content-Disposition: form-data; name="executeActionDTO"\r
-                \r
-                {"actionId":"63285a3388e48972c7519b18","viewMode":false,"paramProperties":{"k0":{"datatype": "string"}}}\r
-                --boundary\r
-                Content-Disposition: form-data; name="k0"; filename="blob"\r
-                Content-Type: text/plain\r
-                \r
-                xyz\r
-                --boundary\r
-                Content-Disposition: form-data; name="parameterMap"\r
-                \r
-                {"Input1.text":"k0"}\r
-                --boundary--""";
-
-        MockServerHttpRequest mock = MockServerHttpRequest
-                .method(HttpMethod.POST, URI.create("https://example.com"))
-                .contentType(new MediaType("multipart", "form-data", Map.of("boundary", "boundary")))
-                .body(parameterMapAtLast);
-
-        final Flux<Part> partsFlux = BodyExtractors.toParts()
-                .extract(mock, this.context);
-
-        NewActionServiceCE newActionServiceSpy = spy(newActionService);
-
-        Mono<ActionExecutionResult> actionExecutionResultMono = newActionServiceSpy.executeAction(partsFlux, null, null);
-
-        ActionExecutionResult mockResult = new ActionExecutionResult();
-        mockResult.setIsExecutionSuccess(true);
-        mockResult.setBody("test body");
-        mockResult.setTitle("test title");
-
-        NewAction newAction = new NewAction();
-        newAction.setId("63285a3388e48972c7519b18");
-        doReturn(Mono.just(mockResult)).when(newActionServiceSpy).executeAction(any(), any());
-        doReturn(Mono.just(newAction)).when(newActionServiceSpy).findByBranchNameAndDefaultActionId(any(), any(), any());
-
-
-        StepVerifier
-                .create(actionExecutionResultMono)
-                .assertNext(response -> {
-                    assertNotNull(response);
-                    assertTrue(response.getIsExecutionSuccess());
-                    assertEquals(mockResult.getBody().toString(), response.getBody().toString());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    public void testParsePartsAndGetParamsFlux_withBlobIdentifiers_replacesValueInParam() {
-        String partsWithBlobRefs = """
-                --boundary\r
-                Content-Disposition: form-data; name="executeActionDTO"\r
-                \r
-                {"actionId":"63285a3388e48972c7519b18","viewMode":false,"paramProperties":{"k0":{"datatype": "string", "blobIdentifiers": ["blob:12345678-1234-1234-1234-123456781234"]}}}\r
-                --boundary\r
-                Content-Disposition: form-data; name="parameterMap"\r
-                \r
-                {"Input1.text":"k0"}\r
-                --boundary\r
-                Content-Disposition: form-data; name="k0"; filename="blob"\r
-                Content-Type: text/plain\r
-                \r
-                {"name": "randomName", "data": "blob:12345678-1234-1234-1234-123456781234"}\r
-                --boundary\r
-                Content-Disposition: form-data; name="blob:12345678-1234-1234-1234-123456781234"; filename="blob"\r
-                Content-Type: text/plain\r
-                \r
-                xy\\nz\r
-                --boundary--""";
-
-        MockServerHttpRequest mock = MockServerHttpRequest
-                .method(HttpMethod.POST, URI.create("https://example.com"))
-                .contentType(new MediaType("multipart", "form-data", Map.of("boundary", "boundary")))
-                .body(partsWithBlobRefs);
-
-        final Flux<Part> partsFlux = BodyExtractors.toParts()
-                .extract(mock, this.context);
-
-        AtomicLong atomicLong = new AtomicLong();
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        Mono<List<Param>> paramsListMono = newActionService.parsePartsAndGetParamsFlux(partsFlux, atomicLong, executeActionDTO).collectList().cache();
-
-        StepVerifier.create(paramsListMono)
-                .assertNext(paramsList -> {
-                    assertEquals(1, paramsList.size());
-                    Param param = paramsList.get(0);
-                    assertEquals("{\"name\": \"randomName\", \"data\": \"blob:12345678-1234-1234-1234-123456781234\"}", param.getValue());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    public void testEnrichExecutionParams_withBlobReference_performsSubstitutionCorrectly() {
-        AtomicLong atomicLong = new AtomicLong(45L);
-        ExecuteActionDTO executeActionDTO = new ExecuteActionDTO();
-        executeActionDTO.setActionId("testId");
-        executeActionDTO.setViewMode(false);
-        executeActionDTO.setParamProperties(Map.of("k0", new ParamProperty("string", List.of("blobId"))));
-        executeActionDTO.setParameterMap(Map.of("Input1.text", "k0"));
-        executeActionDTO.setBlobValuesMap(Map.of("blobId", "xy\\nz"));
-        Param param1 = new Param();
-        param1.setValue("{\"name\": \"randomName\", \"data\": \"blobId\"}");
-        param1.setPseudoBindingName("k0");
-        List<Param> params = List.of(param1);
-
-        Mono<ExecuteActionDTO> enrichedDto = newActionService.enrichExecutionParam(atomicLong, executeActionDTO, params);
-
-        StepVerifier.create(enrichedDto)
-                .assertNext(dto -> {
-                    assertEquals(45, dto.getTotalReadableByteCount());
-
-                    Param param = dto.getParams().get(0);
-                    assertEquals(ClientDataType.STRING, param.getClientDataType());
-                    assertEquals("{\"name\": \"randomName\", \"data\": \"xy\\\\nz\"}", param.getValue());
-                })
-                .verifyComplete();
-    }
 }
